+++ src/types.ts
@@ -1,6 +1,12 @@
 export type SupernovaPluginOptions = {
-    apiToken: string
-    workspaceId: string
-    designSystemId: string
-    designSystemVersionId: string
+  apiToken: string
+  workspaceId: string
+  designSystemId: string
+  designSystemVersionId: string
+
+  searchOptions?: {
+    indexText: boolean
+    indexPageTitles: boolean
+    indexGroupTitles: boolean
+  }
 }