+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to this Supernova <> Gatsby data plugin are mentioned here in this file.
 
+## [0.7.1] - 2022-02-04
+### Search index
+
+We have added ability to access search index data. Index is automatically created and its content can be configured through initial configuration:
+
+```typescript
+
+const supernovaConfig = {
+    ...
+    searchOptions: {
+        indexText: true, // Enables indexing of text from all blocks
+        indexPageTitles: true, // Enables indexing of page titles
+        indexGroupTitles: true // Enables indexing of group titles
+    }
+};
+
+module.exports = {
+    plugins: [
+        {
+            resolve: "@supernovaio/gatsby-source-supernova",
+            options: supernovaConfig,
+        }
+    ]
+}
+
+```
+
+
+Note that index is now enabled by default for all builds.
+
+
 ## [0.7.0] - 2022-29-03
 ### Custom exporter data + new data types
 
@@ -13,6 +44,8 @@ We have added ability to access data defined in custom exporter packages. This a
 - New `Tabs` block data model is available in GQL (+ all required subblocks)
 - New `Table` block data model is available in GQL (+ all required subblocks)
 - New `Column` block data model is available in GQL (+ all required subblocks)
+- `DocumentationItem` of type `Page` now contains new property `previousPageId` that links to previous page object (or null)
+- `DocumentationItem` of type `Page` now contains new property `nextPageId` that links to the next page object (or null)
 
 We have additionally added ability to access `variant` of each block set in the editor through `variantKey` available for each block. Finally, properties for custom blocks were completely reworked.